+++ lib/server/rumbleship-context.d.ts
@@ -14,7 +14,7 @@ export interface RumbleshipContextOptionsPlain {
 interface Context {
     id: string;
     beeline: RumbleshipBeeline;
-    trace: HoneycombSpan | null;
+    trace?: HoneycombSpan;
     container: ContainerInstance;
     authorizer: Authorizer;
     logger: SpyglassLogger;
@@ -25,7 +25,7 @@ export declare class RumbleshipContext implements Context {
     logger: SpyglassLogger;
     authorizer: Authorizer;
     beeline: RumbleshipBeeline;
-    trace: null;
+    trace: HoneycombSpan | undefined;
     private static initialized;
     private static _serviceFactories;
     static addSequelizeServicesToContext: (c: RumbleshipContext) => RumbleshipContext;
@@ -34,4 +34,12 @@ export declare class RumbleshipContext implements Context {
     constructor(id: string, container: ContainerInstance, logger: SpyglassLogger, authorizer: Authorizer, beeline: RumbleshipBeeline);
     release(): void;
 }
+export declare function withRumbleshipContext<T>(filename: string, options: RumbleshipContextOptionsPlain, fn: (ctx: RumbleshipContext) => T): Promise<T>;
+/**
+ * Provides a context that has an authorizer and credentials etc specifically for
+ * THIS microservice so it can be used outside of the context of an Http/geaphql request
+ * or GQL subscription.
+ */
+export declare function getRumbleshipContext(filename: string, config: object): RumbleshipContext;
+export declare function releaseRumbleshipContext(context: Context): void;
 export {};